requirements.txt
pip
gunicorn 
numpy
pandas
pickle4
joblib
flask
flask_restful
scikit-learn
tensorflow
protobuf==3.20
Pillow
matplotlib
keras